Rename workspaceStateStore to workspaceStore
