+++ src/App.vue
@@ -11,7 +11,7 @@
 <script setup lang="ts">
 import config from '@/config'
 import { computed, onMounted } from 'vue'
-import { useWorkspaceStore } from '@/stores/workspaceStateStore'
+import { useWorkspaceStore } from '@/stores/workspaceStore'
 import BlockUI from 'primevue/blockui'
 import ProgressSpinner from 'primevue/progressspinner'
 import GlobalDialog from '@/components/dialog/GlobalDialog.vue'
